Added chapterOne exercises plus chapterTwo exercises
#Chapter1: Page 33
IntelliJ IDEA is an Integrated Development Environment from JetBrains, the
creators of the Kotlin language, in which you can write and run Kotlin code.
• IntelliJ IDEA Community edition is a free version to use for the projects in the book.
Kotlin Apprentice Chapter 1: Your Kotlin Development Environment
• Kotlin code runs on many platforms, and one of the most prominent is the Java
Virtual Machine, which will be used for most of the book.
• To build Kotlin projects with IntelliJ IDEA, you need to install the Java Development
Kit, version 8 or above.
• The IntelliJ IDEA app window consists of a number of panels, the most relevant of
which are the Project panel, the Editor panel, and the Run panel.
• The book starter, final and challenge projects can be opened by choosing File ▸ Open
from the IntelliJ IDEA menu and selecting the root folder of the corresponding
project.

#Chapter2: Page 53
Mini-exercises
If you haven’t been following along with the code in IntelliJ IDEA, now’s the time to try
some exercises to test yourself!
1. Declare a constant of type Int called myAge and set it to your age.
Kotlin Apprentice Chapter 2: Expressions, Variables & Constants
2. Declare a variable of type Double called averageAge. Initially, set it to your own age.
Then, set it to the average of your age and my own age of 30.
3. Create a constant called testNumber and initialize it with whatever integer you’d
like. Next, create another constant called evenOdd and set it equal to testNumber
modulo 2. Now change testNumber to various numbers. What do you notice about
evenOdd?
4. Create a variable called answer and initialize it with the value 0. Increment it by 1.
Add 10 to it. Multiply it by 10. Then, shift it to the right by 3. After all of these
operations, what’s the answer?
Key points
• Computers, at their most fundamental level, perform simple mathematics.
• A programming language allows you to write code, which the compiler converts into
instructions that the CPU can execute. Kotlin code on the JVM is first converted to
bytecode.
• Computers operate on numbers in base 2 form, otherwise known as binary.
• Code comments are denoted by a line starting with // or multiple lines bookended
with /* and */.
• Code comments can be used to document your code.
• You can use println to write things to the console area.
• The arithmetic operators are:
Add: +
Subtract: -
Multiply: *
Divide: /
Remainder: %
• Constants and variables give names to data.
• Once you’ve declared a constant, you can’t change its data, but you can change a
variable’s data at any time.
• Always give variables and constants meaningful names to save yourself and your
colleagues headaches later.
• Operators to perform arithmetic and then assign back to the variable:
Add and assign: +=
Subtract and assign: -=
Multiply and assign: *=
Divide and assign: /=

<> Challenges
Before moving on, here are some challenges to test your knowledge of variables and
constants. You can try the code in IntelliJ IDEA to check your answers.
1. Declare a constant exercises with value 9 and a variable exercisesSolved with
value 0. Increment this variable every time you solve an exercise (including this
one).
2. Given the following code:
age = 16
print(age)
age = 30
print(age)
Declare age so that it compiles. Did you use var or val?
3. Consider the following code:
val a: Int = 46
val b: Int = 10
Work out what answer equals when you replace the final line of code above with each of
these options:
// 1
val answer1: Int = (a * 100) + b
// 2
val answer2: Int = (a * 100) + (b * 100)
// 3
val answer3: Int = (a * 100) + (b / 10)
Kotlin Apprentice Chapter 2: Expressions, Variables & Constants
raywenderlich.com 55
4. Add parentheses to the following calculation. The parentheses should show the
order in which the operations are performed and should not alter the result of the
calculation.
5 * 3 - 4 / 2 * 2
5. Declare two constants a and b of type Double and assign both a value. Calculate the
average of a and b and store the result in a constant named average.
6. A temperature expressed in °C can be converted to °F by multiplying by 1.8 then
incrementing by 32. In this challenge, do the reverse: convert a temperature from °F
to °C. Declare a constant named fahrenheit of type Double and assign it a value.
Calculate the corresponding temperature in °C and store the result in a constant
named celcius.
7. Suppose the squares on a chessboard are numbered left to right, top to bottom, with
0 being the top-left square and 63 being the bottom-right square. Rows are
numbered top to bottom, 0 to 7. Columns are numbered left to right, 0 to 7. Declare
a constant position and assign it a value between 0 and 63. Calculate the
corresponding row and column numbers and store the results in constants named
row and column.
8. A circle is made up of 2� radians, corresponding with 360 degrees. Declare a
constant degrees of type Double and assign it an initial value. Calculate the
corresponding angle in radians and store the result in a constant named radians.
9. Declare four constants named x1, y1, x2 and y2 of type Double. These constants
represent the two-dimensional coordinates of two points. Calculate the distance
between these two points and store the result in a constant named distance.