Added chapter four mini exercises and challenges respective of chapter closure.
#Chapter 4 Mini-exercises
#INCOMPLETE

Key points
• You use the Boolean data type Boolean to represent true and false.
• The comparison operators, all of which return a Boolean, are:
Equal: `==`
Not equal: `!=`
Less than: `<`
Greater than: `>`
Less than or equal: `<=`
Greater than or equal: `>=`
• You can use Boolean logic with && and || to combine comparison conditions.
• You use if expressions to make simple decisions based on a condition, and return a
value.
• You use else and else-if within an if expression to extend the decision-making
beyond a single condition.
• You can use a single line if-else expression to make your code more clear and
concise.
• Short circuiting ensures that only the minimal required parts of a Boolean
expression are evaluated.
• Variables and constants belong to a certain scope, beyond which you cannot use
them. A scope inherits visible variables and constants from its parent.
• while loops allow you to perform a certain task a number of times until a condition is
met.
• The break statement lets you break out of a loop.

Challenges
1. What’s wrong with the following code?
val firstName = "Dick"
if (firstName == "Dick") {
 val lastName = "Lucas"
} else if (firstName == "Ray") {
 val lastName = "Wenderlich"
}
val fullName = firstName + " " + lastName
2. In each of the following statements, what is the value of the Boolean answer
constant?
val answer1 = true && true
val answer2 = false || false
val answer3 = (true && 1 != 2) || (4 > 3 && 100 < 1)
val answer4 = ((10 / 2) > 3) && ((10 % 2) == 0)
3. Suppose the squares on a chessboard are numbered left to right, top to bottom, with
0 being the top-left square and 63 being the bottom-right square. Rows are
numbered top to bottom, 0 to 7. Columns are numbered left to right, 0 to 7. Given a
current position on the chessboard, expressed as a row and column number,
calculate the next position on the chessboard, again expressed as a row and column
number. The ordering is determined by the numbering from 0 to 63. The position
after 63 is again 0.
4. Given the coefficients a, b and c, calculate the solutions to a quadratic equation
with these coefficients. Take into account the different number of solutions (0, 1 or
2). If you need a math refresher, this Wikipedia article on the quadratic equation
will help: https://en.wikipedia.org/wiki/Quadratic_formula.
5. Given a month (represented with a String in all lowercase) and the current year
(represented with an Int), calculate the number of days in the month. Remember
that because of leap years, "february" has 29 days when the year is a multiple of 4
but not a multiple of 100. February also has 29 days when the year is a multiple of
400.
6. Given a number as a Double value, determine if this number is a power of 2. (Hint:
you can use log2(number) to find the base 2 logarithm of number. log2(number) will
return a whole number if number is a power of two. You can also solve the problem
using a loop and no logarithm.)
7. Print a table of the first 10 powers of 2.
8. Given a number n, calculate the n-th Fibonacci number. (Recall Fibonacci is 1, 1, 2,
3, 5, 8, 13, ... Start with 1 and 1 and add these values together to get the next value.
The next value is the sum of the previous two. So the next value in this case is 8+13
= 21.)
9. Given a number n, calculate the factorial of n. (Example: 4 factorial is equal to 1 * 2
* 3 * 4.)
10. Given a number between 2 and 12, calculate the odds of rolling this number using
two six-sided dice. Compute it by exhaustively looping through all of the
combinations and counting the fraction of outcomes that give you that value. Don’t
use a formula