Added the chapter seven, and finale from Section One (Section I), challenges respective of the chapter closure.
#Chapter 6 Functions
#COMPLETE

Mini-exercises
N/A

Key points
 null represents the absence of a value.
• Non-null variables and constants must always have a non-null value.
• Nullable variables and constants are like boxes that can contain a value or be empty
(null).
• To work with the value inside a nullable, you must typically first check that the value
is not null.
• The safest ways to work with a nullable’s value is by using safe calls or the Elvis
operator. Use not-null assertions only when appropriate, as they could produce a
runtime error.

Challenges page 112
Challenge 1: It’s prime time
When I’m acquainting myself with a programming language, one of the first things I do
is write a function to determine whether or not a number is prime. That’s your first
challenge.
First, write the following function:
fun isNumberDivisible(number: Int, divisor: Int): Boolean
You’ll use this to determine if one number is divisible by another. It should return true
when number is divisible by divisor.
Hint: You can use the modulo (%) operator to help you out here.
Next, write the main function:
fun isPrime(number: Int): Boolean
This should return true if number is prime, and false otherwise. A number is prime if
it’s only divisible by 1 and itself. You should loop through the numbers from 1 to the
number and find the number’s divisors. If it has any divisors other than 1 and itself,
then the number isn’t prime. You’ll need to use the isNumberDivisible() function you
wrote earlier.
Use this function to check the following cases:
isPrime(6) // false
isPrime(13) // true
isPrime(8893) // true
Hint 1: Numbers less than 0 should not be considered prime. Check for this case at the
start of the function and return early if the number is less than 0.
Hint 2: Use a for loop to find divisors. If you start at 2 and end before the number itself,
then as soon as you find a divisor, you can return false.
Hint 3: If you want to get really clever, you can simply loop from 2 until you reach the
square root of number, rather than going all the way up to number itself. I’ll leave it as an
exercise for you to figure out why. It may help to think of the number 16, whose square
root is 4. The divisors of 16 are 1, 2, 4, 8 and 16.

Challenge 1: You be the compiler
Which of the following are valid statements?
var name: String? = "Ray"
var age: Int = null
val distance: Float = 26.7
var middleName: String? = null

Challenge 2: Divide and conquer
First, create a function that returns the number of times an integer can be divided by
another integer without a remainder. The function should return null if the division
doesn’t produce a whole number. Name the function divideIfWhole.
Then, write code that tries to extract the nullable result of the function. There should
be two cases: upon success, print "Yep, it divides $answer times", and upon failure,
print "Not divisible :[".
Finally, test your function:
1. Divide 10 by 2. This should print `"Yep, it divides 5 times"`
2. Divide 10 by 3. This should print `"Not divisible :["`
Hint 1: Use the following as the start of the function signature:
fun divideIfWhole(value: Int, divisor: Int)
You’ll need to add the return type, which will be a nullable!
Hint 2: You can use the modulo operator (%) to determine if a value is divisible by
another; recall that this operation returns the remainder from the division of two
numbers. For example, 10 % 2 = 0 means that 10 is divisible by 2 with no remainder,
whereas 10 % 3 = 1 means that 10 is divisible by 3 with a remainder of 1.

Challenge 3: Refactor and reduce
The code you wrote in the last challenge used if statements. In this challenge, refactor
that code to use the Elvis operator instead. This time, make it print "It divides X
times" in all cases, but if the division doesn’t result in a whole number, then X should
be 0.