Added the chapter nine from Section Two (Section II), challenges respective of the chapter closure.
#Chapter 9 Maps and Sets
#COMPLETE

Mini-exercises
N/A

Key points
------------| Maps |
• A map is an unordered collection of key-value pairs.
• The keys of a map are all of the same type, and the values are all of the same type.
• Use indexing to get values and to add, update or remove pairs.
• If a key is not in a map, lookup returns null.
• Built-in Kotlin types such as String, Int, Double have efficient hash values out of the
box.
• Use HashMap<K, V> for performance critical code.
------------| Sets |
• Sets are unordered collections of unique values of the same type.
• Sets are most useful when you need to know whether something is included in the
collection or not.

Challenges page 139
1. Which of the following are valid statements?
1. val map1: Map<Int to Int> = emptyMap()
2. val map2 = emptyMap()
3. val map3: Map<Int, Int> = emptyMap()
For the next four statements, use the following map:
val map4 = mapOf("One" to 1, "Two" to 2, "Three" to 3)
4. map4[1]
5. map4["One"]
6. map4["Zero"] = 0
7. map4[0] = "Zero"
For the next three statements, use the following map:
val map5 = mutableMapOf("NY" to "New York", "CA" to "California")
8. map5["NY"]
9. map5["WA"] = "Washington"
10. map5["CA"] = null
2. Given a map with two-letter state codes as keys, and the full state names as values,
write a function that prints all the states with names longer than eight characters.
For example, for the map mapOf("NY" to "New York", "CA" to "California"), the
output would be California.
3. Write a function that combines two maps into one. If a certain key appears in both
maps, ignore the pair from the first maps. This is the function’s signature:
fun mergeMaps(map1: Map<String, String>, map2: Map<String, String>):
Map<String, String>
4. Declare a function occurrencesOfCharacters that calculates which characters occur
in a string, as well as how often each of these characters occur. Return the result as
a map. This is the function signature:
fun occurrencesOfCharacters(text: String): Map<Char, Int>
Hint: String is a collection of characters that you can iterate over with a for statement.
Bonus: To make your code shorter, maps have a special function that lets you add a
default value if it is not found in the map. For example, map.getOrDefault('a',
defaultValue = 0) returns 0 for the character 'a' if it is not found, instead of simply
returning null.
5. Write a function that returns true if all of the values of a map are unique. Use a set
to test uniqueness. This is the function signature:
fun isInvertible(map: Map<String, Int>): Boolean
6. Given the map:
val nameTitleLookup: Map<String, String?>
 = mutableMapOf("Mary" to "Engineer", "Patrick" to "Intern", "Ray" to
"Hacker")
Set the value of the key "Patrick" to null and completely remove the key and value for
"Ray".