Added the chapter Ten from Section Two (Section II), challenges respective of the chapter closure.
#Chapter 10 Lambdas
#COMPLETE

Mini-exercises
1. Create a constant list called nameList which contains some names as strings. Any
names will do — make sure there’s more than three. Now use fold to create a string
which is the concatenation of each name in the list.
2. Using the same nameList list, first filter the list to contain only names which have
more than four characters in them, and then create the same concatenation of
names as in the above exercise. (Hint: you can chain these operations together.)
3. Create a constant map called namesAndAges which contains some names as strings
mapped to ages as integers. Now use filter to create a map containing only people
under the age of 18.
4. Using the same namesAndAges map, filter out the adults (those 18 or older) and then
use map to convert to a list containing just the names (i.e., drop the ages).

Key points
------------| Lambdas |
Key points
• Lambdas are functions without names. They can be assigned to variables and passed
as arguments to functions.
• Lambdas have shorthand syntax that makes them a lot easier to use than other
functions.
• A lambda can capture the variables and constants from its surrounding context.
• A lambda can be used to direct how a collection is sorted.
• There exists a handy set of functions on collections which can be used to iterate over
the collection and transform the collection. Transforms include mapping each
element to a new value, filtering out certain values, and folding or reducing the
collection down to a single value.

Challenges page 162
Challenge 1: Repeating yourself
Your first challenge is to write a function that will run a given lambda a given number of
times.
Declare the function like so:
fun repeatTask(times: Int, task: () -> Unit)
The function should run the task lambda times number of times.
Use this function to print "Kotlin Apprentice is a great book!" 10 times.
Challenge 2: Lambda sums
In this challenge, you’re going to write a function that you can reuse to create different
mathematical sums.
Declare the function like so:
fun mathSum(length: Int, series: (Int) -> Int) -> Int
The first parameter, length, defines the number of values to sum. The second
parameter, series, is a lambda that can be used to generate a series of values. series
should have a parameter that is the position of the value in the series and return the
value at that position.
mathSum should calculate length number of values, starting at position 1, and return
their sum.
Use the function to find the sum of the first 10 square numbers, which equals 385. Then
use the function to find the sum of the first 10 Fibonacci numbers, which equals 143.
For the Fibonacci numbers, you can use the function you wrote in the challenges of the
functions chapter — or grab it from the solutions if you’re unsure what you’ve done is
correct.
Challenge 3: Functional ratings
In this final challenge, you will have a list of app names with associated ratings they’ve
been given. Note — these are all fictional apps!
Create the data map like so:
val appRatings = mapOf(
 "Calendar Pro" to arrayOf(1, 5, 5, 4, 2, 1, 5, 4),
 "The Messenger" to arrayOf(5, 4, 2, 5, 4, 1, 1, 2),
 "Socialise" to arrayOf(2, 1, 2, 2, 1, 2, 4, 2)
)
First, create a map called averageRatings which will contain a mapping of app names to
average ratings. Use forEach to iterate through the appRatings map, then use reduce to
calculate the average rating and store this rating in the averageRatings map.
Finally, use filter and map chained together to get a list of the app names whose
average rating is greater than 3.