Added the chapter Eleven from Section Three (Section III: Building your own types) and the challenges respective of the chapter closure.
#Chapter 11 Classes
#COMPLETE

Mini-exercises
1. Change the value of lastName on homeOwner, then try reading fullName on both john and
   homeOwner. What do you observe?
2. Write a function memberOf(person: Person, group: [Person]) -> Bool that will
   return true if person can be found inside group, and false if it can not.
   Test it by creating two arrays of five Person objects for group and using john as the
   person. Put john in one of the arrays, but not in the other.
3. Add a property with a custom getter to Student that returns the student’s Grade Point
   Average, or GPA. A GPA is defined as the number of points earned divided by the
   number of credits taken. For the example above, Jane earned (9 + 16 = 25) points while
   taking (3 + 4 = 7) credits, making her GPA (25 / 7 = 3.57).
   Note: Points in most American universities range from 4 per credit for an A, down
   to 1 point for a D (with an F being 0 points). For this exercise, you may of course
   use any scale that you want!

Key points
------------| Classes |

• Classes are a named type that can have properties and methods.
• Classes use references that are shared on assignment.
• Class instances are called objects.
• Objects are mutable.
• Mutability introduces state, which adds complexity when managing your objects.
• Data classes allow you to create simple model objects that avoid a lot of boilerplate
for comparing, printing, and copying objects.
• Destructuring declarations allow you to easily extract multiple properties of data
class objects.

Challenges page 162
Challenge 1: Movie lists
Imagine you’re writing a movie-viewing application in Kotlin. Users can create lists of
movies and share those lists with other users.
Create a User class and a MovieList class that maintains lists for users.
• User: Has a method addList() which adds the given list to a mutable map of
MovieList objects (using the name as a key), and list(): MovieList? which will
return the MovieList for the provided name.
• MovieList: Contains a name and a mutable list of movie titles. A print method will
print all the movies in the movie list.
• Create jane and john users and have them create and share lists. Have both jane and
john modify the same list and call print from both users. Are all the changes
reflected?
Challenge 2: T-Shirt store — data classes
Your challenge here is to build a set of objects to support a T-shirt store. Decide if each
object should be a class or a data class, and go ahead and implement them all.
• TShirt: Represents a shirt style you can buy. Each TShirt has a size, color, price, and
an optional image on the front.
• User: A registered user of the t-shirt store app. A user has a name, email, and a
ShoppingCart (see below).
• Address: Represents a shipping address, containing the name, street, city, and zip
code.
• ShoppingCart: Holds a current order, which is composed of a list of TShirts that the
User wants to buy, as well as a method to calculate the total cost. Additionally, there
is an Address that represents where the order will be shipped.